Created dtos for all component domain models.

+++ backend/KAZABUILD.Domain/Entities/Components/Components/GPUComponent.cs
@@ -19,7 +19,7 @@ namespace KAZABUILD.Domain.Entities.Components.Components
         /// </summary>
         [Required]
         [Range(256, 262144, ErrorMessage = "Video Memory must be between 256 MB and 262144 MB (256 GB)")]
-        public int VideoMemoryAmount { get; set; } = default!;
+        public decimal VideoMemoryAmount { get; set; } = default!;
 
         /// <summary>
         /// What type of Video Memory the GPU uses (e.g., GDDR6, GDDR6X, HBM2).
@@ -33,14 +33,14 @@ namespace KAZABUILD.Domain.Entities.Components.Components
         /// </summary>
         [Required]
         [Range(100, 8000, ErrorMessage = "Core Base Clock Speed must be between 100 MHz and 8000 MHz")]
-        public int CoreBaseClockSpeed { get; set; } = default!;
+        public decimal CoreBaseClockSpeed { get; set; } = default!;
 
         /// <summary>
         /// GPU boost core clock speed in MHz.
         /// </summary>
         [Required]
         [Range(100, 8000, ErrorMessage = "Core Boost Clock Speed must be between 100 MHz and 8000 MHz")]
-        public int CoreBoostClockSpeed { get; set; } = default!;
+        public decimal CoreBoostClockSpeed { get; set; } = default!;
 
         /// <summary>
         /// Number of specialized GPU cores/shaders for task division.
@@ -76,14 +76,14 @@ namespace KAZABUILD.Domain.Entities.Components.Components
         /// </summary>
         [Required]
         [Range(10, 600, ErrorMessage = "GPU length must be between 10 and 600 mm")]
-        public int Length { get; set; } = default!;
+        public decimal Length { get; set; } = default!;
 
         /// <summary>
         /// Thermal Design Power (TDP), the maximum heat the CPU can generate in Watts.
         /// </summary>
         [Required]
         [Range(1, 2000, ErrorMessage = "Thermal design power must be between 1 and 2000 W")]
-        public int ThermalDesignPower { get; set; } = default!;
+        public decimal ThermalDesignPower { get; set; } = default!;
 
         /// <summary>
         /// Number of case expansion slots the GPU occupies.
