+++ backend/KAZABUILD.Domain/Entities/Components/Components/MonitorComponent.cs
@@ -12,7 +12,7 @@ namespace KAZABUILD.Domain.Entities.Components.Components
         /// </summary>
         [Required]
         [Range(1, 200, ErrorMessage = "Screen Size must be between 1 and 200 inches")]
-        public int ScreenSize { get; set; } = default!;
+        public decimal ScreenSize { get; set; } = default!;
 
         /// <summary>
         /// Horizontal Resolution in pixels.
@@ -33,7 +33,7 @@ namespace KAZABUILD.Domain.Entities.Components.Components
         /// </summary>
         [Required]
         [Range(5, 2000, ErrorMessage = "Refresh Rate must be between 5 and 2000 Hz")]
-        public int MaxRefreshRate { get; set; } = default!;
+        public decimal MaxRefreshRate { get; set; } = default!;
 
         /// <summary>
         /// Type of technology used to generate images on the monitor (e.g., IPS, VA, TN, OLED).
@@ -48,7 +48,7 @@ namespace KAZABUILD.Domain.Entities.Components.Components
         /// </summary>
         [Required]
         [Range(0, 200, ErrorMessage = "Response Time must be between 0 and 200 ms")]
-        public int ResponseTime { get; set; } = default!;
+        public decimal ResponseTime { get; set; } = default!;
 
         /// <summary>
         /// Maximum Viewing Angle where the screen is still visible to a human.
